testing total population with new total_population file
